refactor: 인터페이스를 활용하여 트랜잭션 서비스 추상화
