+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -11,7 +11,7 @@ import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class UserServiceTest {
@@ -32,7 +32,7 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao);
+        final var userService = new AppUserService(userDao, userHistoryDao);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -47,20 +47,26 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var userService = new TxUserService(appUserService);
 
         final var oldPassword = userService.findById(1L).getPassword();
         final var newPassword = "newPassword";
         final var createBy = "gugu";
+
         // 트랜잭션이 정상 동작하는지 확인하기 위해 의도적으로 MockUserHistoryDao에서 예외를 발생시킨다.
         assertThrows(DataAccessException.class,
                 () -> userService.changePassword(1L, newPassword, createBy));
 
         final var actual = userService.findById(1L);
 
-        assertAll(
-                () -> assertThat(actual.getPassword()).isNotEqualTo(newPassword),
-                () -> assertThat(actual.getPassword()).isEqualTo(oldPassword)
+        assertSoftly(
+                softly -> {
+                    softly.assertThat(actual.getPassword()).isNotEqualTo(newPassword);
+                    softly.assertThat(actual.getPassword()).isEqualTo(oldPassword);
+                }
         );
     }
 }
